# src/api/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CCDC-style competition scoring engine. Manages scoring data and serves check assignments to agents.",
        "title": "Quincy API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "description": "Returns every user in every userlist for every team, including current passwords. Shape: {\"team\": {\"userlist\": [User]}}.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all scoring users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/types.User"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "description": "Updates the password for a scoring user in the specified userlist for the specified team.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Submit a password change request (PCR)",
                "parameters": [
                    {
                        "description": "Password change request",
                        "name": "pcr",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.PCR"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "types.User": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string",
                    "example": "quin.cy"
                },
                "netbios": {
                    "type": "string",
                    "example": "QUIN"
                },
                "password": {
                    "type": "string",
                    "example": "BuyMyNFT1!"
                },
                "username": {
                    "type": "string",
                    "example": "geraldo"
                }
            }
        },
        "users.PCR": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string",
                    "example": "quin.cy"
                },
                "netbios": {
                    "type": "string",
                    "example": "QUIN"
                },
                "password": {
                    "type": "string",
                    "example": "BuyMyNFT1!"
                },
                "team_num": {
                    "type": "integer",
                    "example": 1
                },
                "user_list": {
                    "type": "string",
                    "example": "local users"
                },
                "username": {
                    "type": "string",
                    "example": "geraldo"
                }
            }
        }
    }
}